requirements.txt
Django>=5.0
Pillow>=10.0
mysqlclient>=2.1
djangorestframework>=3.14
gunicorn
whitenoise
dj-database-url
